Fix AbstractHandlersFactory
