+++ src/main/java/it/pagopa/pn/papertracker/service/handler_step/AR/FinalEventBuilderAr.java
@@ -9,7 +9,7 @@ import it.pagopa.pn.papertracker.middleware.dao.PaperTrackingsDAO;
 import it.pagopa.pn.papertracker.middleware.dao.dynamo.entity.*;
 import it.pagopa.pn.papertracker.middleware.msclient.DataVaultClient;
 import it.pagopa.pn.papertracker.model.HandlerContext;
-import it.pagopa.pn.papertracker.service.handler_step.GenericFinalEventBuilder;
+import it.pagopa.pn.papertracker.service.handler_step.generic.GenericFinalEventBuilder;
 import it.pagopa.pn.papertracker.service.handler_step.HandlerStep;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
