+++ src/main/java/it/pagopa/pn/papertracker/service/handler_step/RIR/FinalEventBuilderRir.java
@@ -2,7 +2,7 @@ package it.pagopa.pn.papertracker.service.handler_step.RIR;
 
 import it.pagopa.pn.papertracker.middleware.dao.PaperTrackingsDAO;
 import it.pagopa.pn.papertracker.middleware.msclient.DataVaultClient;
-import it.pagopa.pn.papertracker.service.handler_step.GenericFinalEventBuilder;
+import it.pagopa.pn.papertracker.service.handler_step.generic.GenericFinalEventBuilder;
 import it.pagopa.pn.papertracker.service.handler_step.HandlerStep;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
