+++ src/main/java/it/pagopa/pn/papertracker/service/handler_step/RIR/SequenceValidatorRir.java
@@ -2,7 +2,7 @@ package it.pagopa.pn.papertracker.service.handler_step.RIR;
 
 import it.pagopa.pn.papertracker.config.SequenceConfiguration;
 import it.pagopa.pn.papertracker.middleware.dao.PaperTrackingsDAO;
-import it.pagopa.pn.papertracker.service.handler_step.GenericSequenceValidator;
+import it.pagopa.pn.papertracker.service.handler_step.generic.GenericSequenceValidator;
 import it.pagopa.pn.papertracker.service.handler_step.HandlerStep;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
